Backport r1187156.
Take qint64's instead of int's in KLimitedIODevice's constructor.

The class attributes were already qint64, and receiving int's made it fail on huge (2GB+) files.

CCBUG: 216672

svn path=/branches/KDE/4.5/kdelibs/; revision=1187161

+++ kdecore/io/klimitediodevice.cpp
@@ -18,7 +18,7 @@
 
 #include "klimitediodevice_p.h"
 
-KLimitedIODevice::KLimitedIODevice( QIODevice *dev, int start, int length )
+KLimitedIODevice::KLimitedIODevice( QIODevice *dev, qint64 start, qint64 length )
     : m_dev( dev ), m_start( start ), m_length( length )
 {
     //kDebug(7005) << "start=" << start << "length=" << length;
